+++ Azure Resource Manager/SharePoint-ADFS/dsc/ConfigureDCVM.ps1
@@ -16,7 +16,7 @@
     Import-DscResource -ModuleName NetworkingDsc -ModuleVersion 9.1.0
     Import-DscResource -ModuleName ActiveDirectoryCSDsc -ModuleVersion 5.0.0
     Import-DscResource -ModuleName CertificateDsc -ModuleVersion 6.0.0
-    Import-DscResource -ModuleName DnsServerDsc -ModuleVersion 3.0.0
+    Import-DscResource -ModuleName DnsServerDsc -ModuleVersion 3.0.1
     Import-DscResource -ModuleName ComputerManagementDsc -ModuleVersion 10.0.0 # Custom
     Import-DscResource -ModuleName AdfsDsc -ModuleVersion 1.4.0